benchling-mcp-server/pyproject.toml
[project]
name = "benchling-mcp-server"
version = "0.1.0"
description = "An MCP server that connects to the Benchling Warehouse."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Benchling" }]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "benchling-sdk>=1.21.2",
    "mcp[cli]>=1.5.0",
    "psutil>=5.9.8",
    "psycopg2-binary>=2.9.10",
    "pypdf>=4.0.0"
]

[project.optional-dependencies]
"literature-search" = [
    "httpx>=0.27.0",
    "pypdf>=4.0.0",
]

[project.scripts]
benchling-mcp-server = "benchling_mcp_server.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.6.0",
    "pytest-asyncio>=0.26.0",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_decorators = false

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004"]  # Ignore magic values in tests

[tool.ruff.lint.isort]
known-first-party = ["benchling_mcp_server"]
known-third-party = ["benchling_sdk", "mcp", "psycopg2"]
